average_screen_color/requirements.txt
mss==4.0.3
Pillow==6.2.1
